Add CI
Adds KEEP_GENERATED env var to retain the files
generated from tests.

+++ tests/test_cli.py
@@ -9,12 +9,15 @@ from unittest_expander import foreach, expand
 
 from py2many.cli import main, _get_all_settings
 
+KEEP_GENERATED = os.environ.get("KEEP_GENERATED", False)
+
 
 @expand
 class CodeGeneratorTests(unittest.TestCase):
     TEST_CASES = ["fib", "rect", "infer", "infer-ops", "binit", "lambda", "int_enum"]
     TESTS_DIR = Path(__file__).parent
     SETTINGS = _get_all_settings(Mock(indent=4))
+    maxDiff = None
 
     def setUp(self):
         os.chdir(self.TESTS_DIR)
@@ -37,7 +40,8 @@ class CodeGeneratorTests(unittest.TestCase):
                     self.assertEqual(f2.read(), actual.read())
         finally:
             try:
-                os.unlink(f"cases/{case}{ext}")
+                if not KEEP_GENERATED:
+                    os.unlink(f"cases/{case}{ext}")
             except FileNotFoundError:
                 pass
